Cast slopeTo result to double

+++ collinear/Point.java
@@ -62,7 +62,7 @@ public class Point implements Comparable<Point> {
         if (this.y == that.y && this.x == that.x) return Double.NEGATIVE_INFINITY;
         else if (this.y == that.y) return +0.0;
         else if (this.x == that.x) return Double.POSITIVE_INFINITY;
-        else return (that.y - this.y) / (that.x - this.x);
+        else return (double) (that.y - this.y) / (that.x - this.x);
     }
 
     /**
